feat(values): share exact numeric semantics
