+++ crates/planner/Cargo.toml
@@ -5,6 +5,7 @@ edition = "2024"
 
 [dependencies]
 helix-ast = { path = "../ast" }
+helix-value-semantics = { path = "../value-semantics" }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 thiserror = "2"
