+++ crates/planner/src/ir/index/range/value.rs
@@ -1,6 +1,7 @@
 use std::cmp::Ordering;
 
 use helix_ast::value::PropertyValue;
+use helix_value_semantics::CanonicalNumber;
 use serde::de::Error as DeError;
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 
@@ -201,10 +202,37 @@ impl RangeIndexLiteral {
 
     pub(super) fn partial_cmp_same_type(&self, other: &Self) -> Option<Ordering> {
         match (self, other) {
-            (Self::I64(left), Self::I64(right)) => Some(left.cmp(right)),
+            (
+                Self::I64(_) | Self::F64(_) | Self::F32(_),
+                Self::I64(_) | Self::F64(_) | Self::F32(_),
+            ) => {
+                let left = match self {
+                    Self::I64(value) => CanonicalNumber::from_i64(*value),
+                    Self::F64(value) => {
+                        CanonicalNumber::from_f64(value.get()).expect("range f64 excludes NaN")
+                    }
+                    Self::F32(value) => {
+                        CanonicalNumber::from_f32(value.get()).expect("range f32 excludes NaN")
+                    }
+                    Self::DateTime(_) | Self::String(_) => {
+                        unreachable!("numeric range arm contains only numbers")
+                    }
+                };
+                let right = match other {
+                    Self::I64(value) => CanonicalNumber::from_i64(*value),
+                    Self::F64(value) => {
+                        CanonicalNumber::from_f64(value.get()).expect("range f64 excludes NaN")
+                    }
+                    Self::F32(value) => {
+                        CanonicalNumber::from_f32(value.get()).expect("range f32 excludes NaN")
+                    }
+                    Self::DateTime(_) | Self::String(_) => {
+                        unreachable!("numeric range arm contains only numbers")
+                    }
+                };
+                Some(left.cmp(&right))
+            }
             (Self::DateTime(left), Self::DateTime(right)) => Some(left.cmp(right)),
-            (Self::F64(left), Self::F64(right)) => left.value.partial_cmp(&right.value),
-            (Self::F32(left), Self::F32(right)) => left.value.partial_cmp(&right.value),
             (Self::String(left), Self::String(right)) => Some(left.cmp(right)),
             (
                 Self::I64(_) | Self::DateTime(_) | Self::F64(_) | Self::F32(_) | Self::String(_),
