+++ crates/value-semantics/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "helix-value-semantics"
+version = "0.1.0"
+edition = "2024"
+
+[lints]
+workspace = true
